Add a Chrome badge blacklist
Add a blacklist of URLs that the Chrome extension's badge will never
show a number for.

This is done by adding a separate API for the Chrome extension to use to
get its badge number, and this API just returns 0 for blacklisted URLs.
Doing it this way means that the (currently hardcoded) blacklist is
server-side, so we can move it into the database and then add an admin
page for modifying it.