Use deform-generated forms for claim/groups
This commit removes custom form code from the templates for account
claim and groups packages, in favour of using generated form code from
deform, with the templates added in the previous commit.

The resulting templates *should* be pixel-for-pixel identical, with the
added bonus of some initial ARIA annotations on form fields, to help
navigating forms with a screenreader.