Render the standalone login form on the server
This commit replaces the client-side rendered login form with one
rendered by deform on the server.

While it would appear that this commit just adds code, that's because
it's only an incremental improvement. It actually removes dependencies
on a bunch of components we want to remove, and will be able to when the
rest of the account forms are rewritten in a similar way. This
changeset:

- sets us on the way to a drastic simplification of the account form
  handling code both on the server and in the client.
- makes a couple of small improvements to user experience (e.g. there is
  no longer a one second pause after logging in)
- prepares the way for doing "login in a popup," which solves some
  difficult Same Origin Policy problems.
- makes it much easier to understand what's going on in the login/logout
  ajax requests sent by the client: the login/logout controller no
  longer uses AjaxFormViewMapper