Replace SQLite with PostgreSQL
Replace SQLite with PostgreSQL for development, testing and on Travis CI.

Use Docker to install PostgreSQL in the development docs.

Also change to using Docker to install Elasticsearch and NSQ in the
development docs.

SQLite behaves differently than PostgreSQL and is a source of dev<->prod
differences and potential bugs.

With Docker installing PostgreSQL is very easy and cross-platform so
there's no reason to use SQLite.

Travis CI also has PostgreSQL built in so there's no reason to use
SQLite.